refactor(cli): extract review-card spawn — phase 5 redo (card e69b55e5)
Same extraction pattern as phases 1 (#1063), 2 (#1064), 3 (#1065): hard-reset to current rust-rewrite tip, re-extract run_review + auto_spawn_review_card + format_review_title into work_commands_review.rs. work_commands_gh.rs's open_pr_and_link uses crate::work_commands_review::auto_spawn_review_card.

work_commands.rs: 1713 → 1560 LOC (-153)
work_commands_review.rs: 0 → ~162 LOC

20 work_commands tests pass; clippy clean (with the #1058-pending pr_create_base_branch allow); fmt clean.

Original #1048 closed due to post-#1049 cross-module call conflict.
